add interface for exporting metadata, refactor
