+++ python/ray/data/_internal/stats.py
@@ -8,7 +8,6 @@ from contextlib import contextmanager
 from dataclasses import dataclass, fields
 from typing import Any, Dict, List, Mapping, Optional, Set, Tuple, Union
 from uuid import uuid4
-import os
 
 import numpy as np
 
@@ -27,12 +26,7 @@ from ray.data.context import DataContext
 from ray.util.annotations import DeveloperAPI
 from ray.util.metrics import Gauge
 from ray.util.scheduling_strategies import NodeAffinitySchedulingStrategy
-
-from ray._private.event.export_event_logger import (
-    EventLogType,
-    get_export_event_logger,
-    check_export_api_enabled,
-)
+from ray.data._internal.metadata_exporter import LoggerDataMetadataExporter
 
 logger = logging.getLogger(__name__)
 
@@ -173,10 +167,8 @@ class _StatsActor:
         # Cache of calls to ray.nodes() to prevent unnecessary network calls
         self._ray_nodes_cache: Dict[str, str] = {}
 
-        (
-            self._export_logger,
-            self._is_data_metadata_export_enabled,
-        ) = self._init_export_logger()
+        # Initialize the metadata exporter
+        self._metadata_exporter = LoggerDataMetadataExporter.create_if_enabled()
 
         # Ray Data dashboard metrics
         # Everything is a gauge because we need to reset all of
@@ -517,7 +509,8 @@ class _StatsActor:
             "dataset_id": dataset_tag,
             "start_time": start_time,
         }
-        self._maybe_export_data_metadata(dataset_metadata)
+        if self._metadata_exporter is not None:
+            self._metadata_exporter.export_dataset_metadata(dataset_metadata)
 
     def update_dataset(self, dataset_tag: str, state: Dict[str, Any]):
         self.datasets[dataset_tag].update(state)
@@ -577,104 +570,6 @@ class _StatsActor:
         if node_ip_tag is not None:
             tags["node_ip"] = node_ip_tag
         return tags
-
-    def _init_export_logger(self) -> tuple[Optional[logging.Logger], bool]:
-        """Initialize the export logger and check if the export API is enabled.
-        Returns:
-            A tuple containing:
-                - The export logger (or None if export API is not enabled).
-                - A boolean indicating if the export API is enabled for data metadata.
-        """
-        # Proto schemas should be imported within the scope of StatsActor to
-        # prevent serialization errors.
-        from ray.core.generated.export_event_pb2 import ExportEvent
-
-        is_data_metadata_export_api_enabled = check_export_api_enabled(
-            ExportEvent.SourceType.EXPORT_DATA_METADATA
-        )
-        if not is_data_metadata_export_api_enabled:
-            return None, False
-
-        log_directory = os.path.join(
-            ray._private.worker._global_node.get_session_dir_path(), "logs"
-        )
-        logger = None
-        try:
-            logger = get_export_event_logger(
-                EventLogType.DATA_METADATA,
-                log_directory,
-            )
-        except Exception:
-            logger.exception(
-                "Unable to initialize the export event logger, so no Data Metadata export "
-                "events will be written."
-            )
-
-        if logger is None:
-            return None, False
-
-        return logger, is_data_metadata_export_api_enabled
-
-    def _maybe_export_data_metadata(self, dataset_metadata: Dict[str, Any]) -> Any:
-        if not self._is_data_metadata_export_enabled:
-            return
-
-        data_metadata_proto = self._dataset_metadata_to_proto(dataset_metadata)
-        self._export_logger.send_event(data_metadata_proto)
-
-    def _dataset_metadata_to_proto(self, dataset_metadata: Dict[str, Any]) -> Any:
-        """Convert the DAG structure dictionary to a protobuf message.
-
-        Args:
-            dag_structure: Dictionary representation of the DAG structure. This
-            is passed as a dictionary to avoid serializing the protobuf message.
-
-        Returns:
-            The protobuf message representing the DAG structure.
-        """
-        from ray.core.generated.export_data_metadata_pb2 import (
-            ExportDataMetadata,
-            OperatorDAG,
-            Operator,
-            SubOperator,
-        )
-
-        # Create the protobuf message
-        data_metadata = ExportDataMetadata()
-        dag = OperatorDAG()
-
-        # Add operators to the DAG
-        dag_structure = dataset_metadata["dag_structure"]
-        if "operators" in dag_structure:
-            for op_dict in dag_structure["operators"]:
-                operator = Operator()
-                operator.name = op_dict.get("name", UNKNOWN)
-                operator.id = op_dict.get("id", UNKNOWN)
-                operator.uuid = op_dict.get("uuid", UNKNOWN)
-
-                # Add input dependencies if they exist
-                if "input_dependencies" in op_dict:
-                    for dep_id in op_dict["input_dependencies"]:
-                        operator.input_dependencies.append(dep_id)
-
-                # Add sub-operators if they exist
-                if "sub_operators" in op_dict:
-                    for sub_op_dict in op_dict["sub_operators"]:
-                        sub_op = SubOperator()
-                        sub_op.name = sub_op_dict.get("name", UNKNOWN)
-                        sub_op.id = sub_op_dict.get("id", UNKNOWN)
-                        operator.sub_operators.append(sub_op)
-
-                # Add the operator to the DAG
-                dag.operators.append(operator)
-
-        # Populate the data metadata proto
-        data_metadata.dag.CopyFrom(dag)
-        data_metadata.job_id = dataset_metadata.get("job_id", None)
-        data_metadata.start_time = dataset_metadata.get("start_time", None)
-        data_metadata.dataset_id = dataset_metadata.get("dataset_id", None)
-
-        return data_metadata
 
 
 # Creating/getting an actor from multiple threads is not safe.
